services/transcription-service/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
ipywidgets
requests
openai-whisper
psycopg2-binary
python-dotenv